requirements.txt
absl-py==0.1.10
alabaster==0.7.10
anaconda-client==1.2.2
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astroid==1.4.9
astropy==2.0.4
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bcrypt==3.1.4
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.10.0
chardet==3.0.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.1
colorama==0.3.9
conda==4.3.16
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.2
decorator==4.0.11
distributed==1.16.3
Django==2.1.7
django-filter==2.1.0
django-frontend==1.8.0
docutils==0.13.1
entrypoints==0.2.2
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==1.0.2
Flask-Cors==3.0.2
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.2
gevent==1.2.1
greenlet==0.4.12
h5py==2.7.1
HeapDict==1.0.0
html5lib==0.9999999
httplib2==0.10.3
hyperlink==18.0.0
idna==2.5
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
jellyfish==0.5.6
Jinja2==2.10
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.1.4
Kivy==1.10.1
Kivy-Garden==0.1.4
lazy-object-proxy==1.2.2
llvmlite==0.22.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.1.2
MeaningCloud-python==1.1.1
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.12
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.37.0
numexpr==2.6.4
numpy==1.14.1
numpydoc==0.6.0
oauthlib==2.0.6
odo==0.5.0
olefile==0.44
opencv-python==3.4.0.12
openpyxl==2.4.7
opensecrets-crpapi==0.2.2
oscpy==0.3.0
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.1
paramiko==2.4.1
partd==0.3.8
passlib==1.7.1
pathlib2==2.2.1
patsy==0.4.1
pdfminer.six==20170720
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pipenv==11.10.4
ply==3.10
prompt-toolkit==1.0.14
protobuf==3.5.1
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.33
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycryptodome==3.6.4
pycurl==7.43.0.2
pyflakes==1.5.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==1.6.4
PyMySQL==0.8.1
PyNaCl==1.2.1
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pyPdf==1.13
PyPDF2==1.26.0
pyproj==1.9.5.1
pyshp==1.2.12
PySocks==1.6.8
pytest==3.0.7
python-congress==0.3.5
python-dateutil==2.6.0
python-oembed==0.2.4
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzipcode3==2.2
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
requests==2.21.0
requests-oauthlib==0.8.0
rope-py3k==0.9.4.post1
roslibpy==0.3.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
scraper==0.1.0
seaborn==0.7.1
selenium==3.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
sshtunnel==0.1.3.1
statsmodels==0.8.0
sympy==1.0
tables==3.4.2
tblib==1.3.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.6
testpath==0.3
textblob==0.15.1
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
tweepy==3.6.0
Twisted==18.7.0
twython==3.6.0
txaio==18.8.1
unicodecsv==0.14.1
urllib3==1.24.2
us==1.0.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.14.1
widgetsnbextension==2.0.0
wrapt==1.10.10
WTForms==2.1
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
zope.interface==4.5.0
